Finally move template logic into an Helper because it is reused for challenge previews
