+++ live/app/templates/assessments/get-challenge.hbs
@@ -11,7 +11,7 @@
       </div>
     {{/unless}}
 
-    {{component _challengeComponentClass
+    {{component (get-challenge-component-class model.challenge)
         challenge=model.challenge
         assessment=model.assessment
         answers=model.answers
